backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic>=2.0.0
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0